requirements.txt
celery==5.2.1
python-telegram-bot==13.9
PyYAML==6.0
